requirements.txt
absl-py
cachetools
certifi
chardet
dateparser
focal-loss-torch
future
grpcio
idna
importlib-metadata
joblib
Markdown
numpy
oauthlib
opencv-contrib-python
opencv-python
Pillow
protobuf
pyasn1
pyasn1-modules
pytesseract
python-dateutil
pytz
regex
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
tensorboard
tensorboard-plugin-wit
threadpoolctl
tqdm
typing-extensions
tzlocal
urllib3
Werkzeug
zipp
mlflow